Set cookbook source & issues url from metadata
With the inevitability of chef supporting source_url and issues_url in the
metadata.rb, Supermarket should set the attributes on cookbook from the latest
version that gets uploaded.

The reason this does not add a source and issue url to CookbookVersion is - if
the URL changes, we would always want the URL to be the most recent one incase
they change. :)

Adjust external_url signature value in specs

Make it an empty string instead of nil.